+++ sdk/resourcemanager/managementgroups/armmanagementgroups/zz_generated_hierarchysettings_client.go
@@ -12,13 +12,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
 	"net/url"
 	"strings"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // HierarchySettingsClient contains the methods for the HierarchySettings group.
@@ -29,8 +30,15 @@ type HierarchySettingsClient struct {
 }
 
 // NewHierarchySettingsClient creates a new instance of HierarchySettingsClient with the specified values.
-func NewHierarchySettingsClient(con *arm.Connection) *HierarchySettingsClient {
-	return &HierarchySettingsClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewHierarchySettingsClient(credential azcore.TokenCredential, options *arm.ClientOptions) *HierarchySettingsClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &HierarchySettingsClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // CreateOrUpdate - Creates or updates the hierarchy settings defined at the Management Group level.