+++ sdk/resourcemanager/managementgroups/armmanagementgroups/zz_generated_managementgroups_client.go
@@ -12,15 +12,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"net/url"
+	"strconv"
+	"strings"
 )
 
 // ManagementGroupsClient contains the methods for the ManagementGroups group.
@@ -31,8 +31,15 @@ type ManagementGroupsClient struct {
 }
 
 // NewManagementGroupsClient creates a new instance of ManagementGroupsClient with the specified values.
-func NewManagementGroupsClient(con *arm.Connection) *ManagementGroupsClient {
-	return &ManagementGroupsClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewManagementGroupsClient(credential azcore.TokenCredential, options *arm.ClientOptions) *ManagementGroupsClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &ManagementGroupsClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // BeginCreateOrUpdate - Create or update a management group. If a management group is already created and a subsequent create request is issued with different