+++ sdk/resourcemanager/managementgroups/armmanagementgroups/zz_generated_managementgroupsapi_client.go
@@ -11,11 +11,12 @@ package armmanagementgroups
 import (
 	"context"
 	"fmt"
-	"net/http"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
 )
 
 // ManagementGroupsAPIClient contains the methods for the ManagementGroupsAPI group.
@@ -26,8 +27,15 @@ type ManagementGroupsAPIClient struct {
 }
 
 // NewManagementGroupsAPIClient creates a new instance of ManagementGroupsAPIClient with the specified values.
-func NewManagementGroupsAPIClient(con *arm.Connection) *ManagementGroupsAPIClient {
-	return &ManagementGroupsAPIClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewManagementGroupsAPIClient(credential azcore.TokenCredential, options *arm.ClientOptions) *ManagementGroupsAPIClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &ManagementGroupsAPIClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // CheckNameAvailability - Checks if the specified management group name is valid and unique