+++ sdk/resourcemanager/managementgroups/armmanagementgroups/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 0.2.0 (2021-10-28)
+### Breaking Changes
+
+- Function `NewEntitiesClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewManagementGroupsAPIClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewManagementGroupSubscriptionsClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewOperationsClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewManagementGroupsClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewHierarchySettingsClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+
+### New Content
+
+
+Total 6 breaking change(s), 0 additive change(s).
+
+
 ## 0.1.1 (Unreleased)
 
 ### Features Added